move generateCsv to a new file as per PR comments
