+++ shared/ui/Stream/ExportPanel.tsx
@@ -3,38 +3,20 @@ import ScrollBox from "./ScrollBox";
 import { CodeStreamState } from "../store";
 import { useAppDispatch, useAppSelector, useDidMount } from "../utilities/hooks";
 import { HostApi } from "../webview-api";
-import { mapFilter } from "../utils";
-import { CodemarkType } from "@codestream/protocols/api";
+
 import { PanelHeader } from "../src/components/PanelHeader";
 import { CreateCodemarkIcons } from "./CreateCodemarkIcons";
-import { createSelector } from "reselect";
+
 import CancelButton from "./CancelButton";
 import { closePanel } from "./actions";
 import Icon from "./Icon";
-import { stringify } from "csv-stringify/browser/esm/sync";
 import copy from "copy-to-clipboard";
-
-const getSearchableCodemarks = createSelector(
-	(state: CodeStreamState) => state.codemarks,
-	codemarksState => {
-		return mapFilter(Object.values(codemarksState), codemark => {
-			if (
-				!codemark.isChangeRequest &&
-				(codemark.type === CodemarkType.Comment || codemark.type === CodemarkType.Issue)
-			) {
-				return codemark;
-			}
-			return;
-		});
-	}
-);
+import { generateCsv } from "./GenerateCsvFunction";
 
 export const ExportPanel = () => {
 	const dispatch = useAppDispatch();
 	const derivedState = useAppSelector((state: CodeStreamState) => {
-		const codemarks = useAppSelector(getSearchableCodemarks);
-
-		return { codemarks, webviewFocused: state.context.hasFocus, repos: state.repos };
+		return { webviewFocused: state.context.hasFocus, repos: state.repos };
 	});
 	const data = generateCsv();
 
@@ -42,64 +24,6 @@ export const ExportPanel = () => {
 		if (derivedState.webviewFocused)
 			HostApi.instance.track("Page Viewed", { "Page Name": "Export" });
 	});
-
-	function generateCsv() {
-		// repo,file,commitSha,location,date,author,id,parentId,type,title,body,assignees
-		if (!derivedState.codemarks.length) return;
-		let output: {}[] = [];
-		derivedState.codemarks.forEach(codemark => {
-			if (!codemark) return;
-			if (codemark.markers) {
-				codemark.markers.map(marker => {
-					if (!marker) return;
-					const location: any = marker.referenceLocations
-						? marker.referenceLocations[marker.referenceLocations.length - 1] || {}
-						: {};
-					const repo = derivedState.repos[marker.repoId];
-					const repoName = repo ? repo.name : "";
-					output.push(
-						{ repo: repoName },
-						{ file: marker.file },
-						{ commitSha: location.commitHash },
-						{ location: location.location ? location.location[0] : "" },
-						{ date: codemark.createdAt },
-						{ author: codemark.creatorId },
-						{ id: codemark.id },
-						{ parentId: codemark.parentPostId },
-						{ type: codemark.type },
-						{ title: codemark.title || codemark.text },
-						{ body: codemark.title ? codemark.text : "" },
-						{ assignees: codemark.assignees }
-					);
-				});
-			}
-		});
-
-		try {
-			const stringData = stringify(output, {
-				header: true,
-				columns: {
-					repo: "repo",
-					file: "file",
-					commitSha: "commitSha",
-					location: "location",
-					date: "date",
-					author: "author",
-					id: "id",
-					parentId: "parentId",
-					type: "type",
-					title: "title",
-					body: "body",
-					assignees: "assignees",
-				},
-			});
-			return stringData;
-		} catch (ex) {
-			console.log(ex);
-		}
-
-		return;
-	}
 
 	return (
 		<div className="panel full-height activity-panel">
